+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Add clj-toolbox.prelude/not-empty?
 * Add files/size and files/children
 * Add clj-toolbox.atoms
+* Add clj-toolbox.shelltools
 
 ## [0.9.0] — 2024-12-30
 * Add property tests using clojure.test.check for streams and base64